Changed to floating point operations in turning - testing needed

+++ client/src/All_Movements.hpp
@@ -60,14 +60,14 @@ class TurnAngle : public Movement {
 	private:
 		uint32_t timeout, prev_t;
 		int32_t start_enc;
-		int32_t angle, start_angle, prev_angle;
+		float angle, start_angle, prev_angle;
 		int32_t speed, prev_speedD, prev_speedP;
 		Motor *mDrive, *mPivot;
 		IMU *imu;
 		bool right_turn;		
 
-		int32_t encoder_angle();
-		int32_t imu_angle();
+		float encoder_angle();
+		float imu_angle();
 };
 
 /* Drive Onto Ramp - (Drive in reverse, validate position between dist to boundary + accelerometer tilt + roll)*/
